restructured api frontend code layout for more readability
